chore: Moving the crop tests to utils with the crop code.
